requirements.txt
flask
flask-dance
python-dotenv
gunicorn
sentry-sdk[flask]==0.7.10